feat: criado classes de entidades manualmente pois o console não fornece suporte a mapeamento via atributos, pois annotation está descontinuado, atualizado script SQL do banco de dados para inserir vendas

+++ api/modules/taxes/dtos/CreateTaxeDTO.php
@@ -6,7 +6,7 @@ class CreateTaxeDTO {
     
     private String $name;
     
-    private Int $percentage;
+    private Float $percentage;
 
     function __construct(Array $data){
         $this->name = $data['name'];
@@ -18,7 +18,7 @@ class CreateTaxeDTO {
         return $this->name;
     }
 
-    public function getPercentage(): Int
+    public function getPercentage(): Float
     {
         return $this->percentage;
     }
